+++ exif/_app1_metadata.py
@@ -6,8 +6,8 @@ import sys
 from fractions import Fraction
 
 from exif._constants import (
-    ATTRIBUTE_ID_MAP, ExifTypes, HEX_PER_BYTE, ERROR_IMG_NO_ATTR, BYTES_PER_IFD_TAG,
-    ATTRIBUTE_NAME_MAP)
+    ATTRIBUTE_ID_MAP, ExifTypes, ERROR_IMG_NO_ATTR, BYTES_PER_IFD_TAG, ATTRIBUTE_NAME_MAP)
+from exif._hex_interface import HexInterface
 from exif._ifd_tag import IfdTag
 
 
@@ -19,44 +19,45 @@ class App1MetaData(object):
         # Get the tag count in the deletion target's IFD section.
         delete_target = self.ifd_tags[ifd_tag.tag]
         section_start_address = delete_target.section_start_address
-        section_tag_count = int(
-            self.segment_hex[section_start_address:section_start_address + 2 * HEX_PER_BYTE], 16)
+        section_tag_count = int(self._segment_hex.read(section_start_address, 2), 16)
 
         # Decrease the tag count by 1 in the file hexadecimal.
-        self.segment_hex = (self.segment_hex[:section_start_address] +
-                            hex(section_tag_count - 1).lstrip('0x').zfill(4) +
-                            self.segment_hex[section_start_address + 2 * HEX_PER_BYTE:])
+        self._segment_hex.modify_number(section_start_address, 2, section_tag_count - 1)
 
         # Parse over the deletion target's IFD section and remove the tag.
-        cursor = section_start_address + 2 * HEX_PER_BYTE
+        cursor = section_start_address + 2
         for tag_index in range(section_tag_count):  # pylint: disable=unused-variable
             tag = IfdTag(
                 self._endianness,
-                self.segment_hex[cursor:cursor + BYTES_PER_IFD_TAG * HEX_PER_BYTE],
+                self._segment_hex.read(cursor, BYTES_PER_IFD_TAG),
                 section_start_address)
             if delete_target == tag:
-                self.segment_hex = (
-                    self.segment_hex[:cursor] +
-                    self.segment_hex[cursor + BYTES_PER_IFD_TAG * HEX_PER_BYTE:])
+                self._segment_hex.delete(cursor, BYTES_PER_IFD_TAG)
             else:
-                cursor += BYTES_PER_IFD_TAG * HEX_PER_BYTE
+                cursor += BYTES_PER_IFD_TAG
 
         #  Pad ending of IFD section to preserve pointers.
-        self.segment_hex = (
-            self.segment_hex[:cursor] + '00' * BYTES_PER_IFD_TAG + self.segment_hex[cursor:])
+        self._segment_hex.insert_null(cursor, BYTES_PER_IFD_TAG)
 
         # Overwrite pointer data with null bytes (if applicable, depending on datatype).
         if delete_target.dtype in [ExifTypes.ASCII, ExifTypes.RATIONAL, ExifTypes.SRATIONAL]:
-            cursor = (0xA + ifd_tag.value_offset) * HEX_PER_BYTE
+            cursor = 0xA + ifd_tag.value_offset
             if delete_target.dtype == ExifTypes.ASCII:
-                null_data = '00' * delete_target.count
-            else:  # SRATIONAL or RATIONAL
-                null_data = '00' * 8 * delete_target.count
-            self.segment_hex = (self.segment_hex[:cursor] + null_data +
-                                self.segment_hex[cursor + len(null_data):])
+                self._segment_hex.wipe(cursor, delete_target.count)
+            else:  # SRATIONAL or RATIONAL (8 Bytes)
+                self._segment_hex.wipe(cursor, delete_target.count * 8)
 
         # Remove tag from parser tag dictionary.
         del self.ifd_tags[ifd_tag.tag]
+
+    def get_segment_hex(self):
+        """Get equivalent APP1 hexadecimal string.
+
+        :returns: segment hexadecimal string
+        :rtype: str
+
+        """
+        return self._segment_hex.get_hex_string()
 
     def get_tag_list(self):
         """Get a list of EXIF tag attributes present in the image objec.
@@ -70,13 +71,13 @@ class App1MetaData(object):
 
     def _read_ascii_tag(self, ifd_tag):
         retval_chars = []
-        cursor = (0xA + ifd_tag.value_offset) * HEX_PER_BYTE
+        cursor = 0xA + ifd_tag.value_offset
 
         # Subtract 1 from IFD member count to ignore null terminator character.
         for member_index in range(ifd_tag.count - 1):  # pylint: disable=unused-variable
-            current_ascii_val = self.segment_hex[cursor:cursor + 1 * HEX_PER_BYTE]
+            current_ascii_val = self._segment_hex.read(cursor, 1)
             retval_chars.append(chr(int(current_ascii_val, 16)))
-            cursor += 1 * HEX_PER_BYTE
+            cursor += 1
 
         return ''.join(retval_chars)
 
@@ -91,7 +92,7 @@ class App1MetaData(object):
 
     def _read_numeric_tag(self, ifd_tag):
         retvals = []
-        cursor = (0xA + ifd_tag.value_offset) * HEX_PER_BYTE
+        cursor = 0xA + ifd_tag.value_offset
         for member_index in range(ifd_tag.count):  # pylint: disable=unused-variable
             # If type is 4-byte or less, value is stored in member (1, 3, 4, 9).
             if ifd_tag.dtype == ExifTypes.BYTE:
@@ -111,14 +112,14 @@ class App1MetaData(object):
             elif ifd_tag.dtype == ExifTypes.RATIONAL:
                 data_format = ifd_tag.endian_prefix + "LL"
                 numerator, denominator = struct.unpack(data_format, binascii.unhexlify(
-                    self.segment_hex[cursor:cursor + 8 * HEX_PER_BYTE]))
-                cursor += 8 * HEX_PER_BYTE
+                    self._segment_hex.read(cursor, 8)))
+                cursor += 8
                 retvals.append(float(numerator) / float(denominator))
             elif ifd_tag.dtype == ExifTypes.SRATIONAL:
                 data_format = ifd_tag.endian_prefix + "ll"
                 numerator, denominator = struct.unpack(data_format, binascii.unhexlify(
-                    self.segment_hex[cursor:cursor + 8 * HEX_PER_BYTE]))
-                cursor += 8 * HEX_PER_BYTE
+                    self._segment_hex.read(cursor, 8)))
+                cursor += 8
                 retvals.append(float(numerator) / float(denominator))
             else:
                 raise RuntimeError("unknown datatype value")
@@ -134,18 +135,16 @@ class App1MetaData(object):
         if len(value) > ifd_tag.count - 1:
             raise ValueError("string must be no longer than original")
 
-        cursor = (0xA + ifd_tag.value_offset) * HEX_PER_BYTE
-        new_segment_hex = self.segment_hex[:cursor]
-        for character in value:
-            new_segment_hex += hex(ord(character)).lstrip('0x')
-            cursor += 1 * HEX_PER_BYTE
-        for _i in range(ifd_tag.count - 1 - len(value)):
-            new_segment_hex += '00'
-            cursor += 1 * HEX_PER_BYTE
+        ascii_offset = 0xA + ifd_tag.value_offset
+        ascii_hex_string = ''
 
-        new_segment_hex += self.segment_hex[cursor:]
+        for character in value:
+            ascii_hex_string += hex(ord(character)).lstrip('0x')
+        for _i in range(ifd_tag.count - 1 - len(value)):
+            ascii_hex_string += '00'
+
         ifd_tag.count = len(value) + 1
-        self.segment_hex = new_segment_hex
+        self._segment_hex.modify_hex(ascii_offset, ascii_hex_string)
 
     def _modify_ifd_tag(self, ifd_tag, value):
         if ifd_tag.dtype == ExifTypes.ASCII:
@@ -161,7 +160,7 @@ class App1MetaData(object):
             assert ifd_tag.count == 1
             value = (value,)
 
-        cursor = (0xA + ifd_tag.value_offset) * HEX_PER_BYTE
+        cursor = 0xA + ifd_tag.value_offset
         for member_index in range(ifd_tag.count):
             # If type is 4-byte or less, value is stored in member (1, 3, 4, 9).
             if ifd_tag.dtype == ExifTypes.BYTE:
@@ -181,9 +180,8 @@ class App1MetaData(object):
                 new_member_hex = binascii.hexlify(new_member_bits)
                 if sys.version_info[0] == 3:  # pragma: no cover
                     new_member_hex = new_member_hex.decode("utf8")
-                self.segment_hex = (self.segment_hex[:cursor] + new_member_hex +
-                                    self.segment_hex[cursor + 8 * HEX_PER_BYTE:])
-                cursor += 8 * HEX_PER_BYTE
+                self._segment_hex.modify_hex(cursor, new_member_hex)
+                cursor += 8
             elif ifd_tag.dtype == ExifTypes.SRATIONAL:
                 fraction = Fraction(value[member_index]).limit_denominator()
                 new_member_bits = struct.pack(
@@ -191,85 +189,83 @@ class App1MetaData(object):
                 new_member_hex = binascii.hexlify(new_member_bits)
                 if sys.version_info[0] == 3:  # pragma: no cover
                     new_member_hex = new_member_hex.decode("utf8")
-                self.segment_hex = (self.segment_hex[:cursor] + new_member_hex +
-                                    self.segment_hex[cursor + 8 * HEX_PER_BYTE:])
-                cursor += 8 * HEX_PER_BYTE
+                self._segment_hex.modify_hex(cursor, new_member_hex)
+                cursor += 8
 
             else:
                 raise RuntimeError("unknown datatype value")
 
     def _unpack_ifd_tags(self):
-        cursor = 0xA * HEX_PER_BYTE
+        cursor = 0xA
         exif_offset = None
         gps_offset = None
 
         # Read the endianness specified by the image.
-        self._endianness = self.segment_hex[cursor:cursor+2*HEX_PER_BYTE]
-        cursor += 2 * HEX_PER_BYTE
+        self._endianness = self._segment_hex.read(cursor, 2)
+        cursor += 2
 
         # Skip over fixed 0x002A bytes.
-        cursor += 2 * HEX_PER_BYTE
+        cursor += 2
 
         # Determine the location of the first IFD section relative to the start of the APP1 section.
-        first_ifd_offset = self.segment_hex[cursor:cursor + 4 * HEX_PER_BYTE]
         # 0xA is IFD section offset from start of APP1.
-        cursor = (0xA + int(first_ifd_offset, 16)) * HEX_PER_BYTE
+        cursor = 0xA + int(self._segment_hex.read(cursor, 4), 16)
 
         # Read each IFD section.
         current_ifd = 0
-        while cursor != 0xA*HEX_PER_BYTE:
+        while cursor != 0xA:
             section_start_address = cursor
-            num_ifd_tags = int(self.segment_hex[cursor:cursor+2*HEX_PER_BYTE], 16)
-            cursor += 2 * HEX_PER_BYTE
+            num_ifd_tags = int(self._segment_hex.read(cursor, 2), 16)
+            cursor += 2
 
             for tag_index in range(num_ifd_tags):  # pylint: disable=unused-variable
                 tag = IfdTag(
                     self._endianness,
-                    self.segment_hex[cursor:cursor + BYTES_PER_IFD_TAG * HEX_PER_BYTE],
+                    self._segment_hex.read(cursor, BYTES_PER_IFD_TAG),
                     section_start_address)
                 self.ifd_tags[tag.tag] = tag
                 if tag.is_exif_pointer():
                     exif_offset = tag.value_offset
                 if tag.is_gps_pointer():
                     gps_offset = tag.value_offset
-                cursor += BYTES_PER_IFD_TAG * HEX_PER_BYTE
+                cursor += BYTES_PER_IFD_TAG
 
-            cursor = (0xA + int(self.segment_hex[cursor:cursor+4*HEX_PER_BYTE], 16)) * HEX_PER_BYTE
+            cursor = 0xA + int(self._segment_hex.read(cursor, 4), 16)
             current_ifd += 1
 
         # If an EXIF section pointer exists, read EXIF IFD tags.
         if exif_offset:
-            cursor = (0xA + exif_offset) * HEX_PER_BYTE
+            cursor = 0xA + exif_offset
             section_start_address = cursor
-            num_ifd_tags = int(self.segment_hex[cursor:cursor + 2 * HEX_PER_BYTE], 16)
-            cursor += 2 * HEX_PER_BYTE
+            num_ifd_tags = int(self._segment_hex.read(cursor, 2), 16)
+            cursor += 2
 
             for tag_index in range(num_ifd_tags):  # pylint: disable=unused-variable
                 tag = IfdTag(
                     self._endianness,
-                    self.segment_hex[cursor:cursor + BYTES_PER_IFD_TAG * HEX_PER_BYTE],
+                    self._segment_hex.read(cursor, BYTES_PER_IFD_TAG),
                     section_start_address)
                 self.ifd_tags[tag.tag] = tag
-                cursor += BYTES_PER_IFD_TAG * HEX_PER_BYTE
+                cursor += BYTES_PER_IFD_TAG
 
         # If an GPS section pointer exists, read GPS IFD tags.
         if gps_offset:
-            cursor = (0xA + gps_offset) * HEX_PER_BYTE
+            cursor = 0xA + gps_offset
             section_start_address = cursor
-            num_ifd_tags = int(self.segment_hex[cursor:cursor + 2 * HEX_PER_BYTE], 16)
-            cursor += 2 * HEX_PER_BYTE
+            num_ifd_tags = int(self._segment_hex.read(cursor, 2), 16)
+            cursor += 2
 
             for tag_index in range(num_ifd_tags):  # pylint: disable=unused-variable
                 tag = IfdTag(
                     self._endianness,
-                    self.segment_hex[cursor:cursor + BYTES_PER_IFD_TAG * HEX_PER_BYTE],
+                    self._segment_hex.read(cursor, BYTES_PER_IFD_TAG),
                     section_start_address)
                 self.ifd_tags[tag.tag] = tag
-                cursor += BYTES_PER_IFD_TAG * HEX_PER_BYTE
+                cursor += BYTES_PER_IFD_TAG
 
     def __init__(self, segment_hex):
         self._endianness = None
-        self.segment_hex = segment_hex
+        self._segment_hex = HexInterface(segment_hex)
         self.ifd_tags = {}
 
         self._unpack_ifd_tags()